#2 rename: 폴더 네이밍을 구체적으로 변경
DraggableMenu -> draggableSearchMenu
+ 폴더네임 카멜케이스 컨벤션에 맞게 소문자로 시작하도록 수정
