+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import DraggableMenu from './_components/DraggableMenu/DraggableMenu';
+import DraggableMenu from './_components/draggableSearchMenu/DraggableMenu';
 
 export default function Landing() {
   return (
